+++ docs/source/developer/api/response.py
@@ -1,0 +1,4 @@
+:mod:`guillotina.response`
+-------------------------
+
+.. automodule:: guillotina.response